+++ src/comment-npm-publish/index.ts
@@ -1,21 +1,5 @@
 // comment-npm-publish/index.ts
 
-import debug from 'debug';
-import { getInput } from '@actions/core';
-
-import { publishCommentAndRemovePrevious } from '../github-api';
-
-const log = debug('github-actions:comment-npm-publish');
-export async function main(): Promise<void> {
-  log('Action start');
-
-  const packageNameAndBetaVersion = getInput('betaPackage');
-  log('Package Name and Version obtained: ', packageNameAndBetaVersion);
-  await publishCommentAndRemovePrevious(
-    `Beta Published - Install Command: \`npm install ${packageNameAndBetaVersion}\` `.replaceAll('"', ''),
-    'Beta Published - Install Command: ',
-  );
-  log('Action end');
-}
+import main from './comment-npm-publish';
 
 await main();
