+++ src/validate-npm-package/index.ts
@@ -1,7 +1,5 @@
 // validate-npm-package/index.ts
 
-import { getInput } from '@actions/core';
-
 import main from './validate-npm-package';
 
-await main(getInput('betaPackage'));
+await main();
